Trait zen_parser::prelude::AsciiRead

pub trait AsciiRead: Read + Seek {

    fn error(&mut self, kind: ErrorCode) -> Error { ... }

    fn byte(&mut self) -> Result<u8, Error> { ... }

    fn peek(&mut self) -> Result<u8, Error> { ... }

    fn peek_tuple(&mut self) -> Result<(u8, u8), Error> { ... }

    fn test_for(&mut self, s: &str) -> Result<bool, Error> { ... }

    fn consume(&mut self, s: &str) -> Result<bool, Error> { ... }

    fn consume_until(&mut self, byte: u8) -> Result<u32, Error> { ... }

    fn consume_whitespaces(&mut self) -> Result<(), Error> { ... }

    fn string_until(&mut self, byte: u8) -> Result<String, Error> { ... }

    fn string_until_whitespace(&mut self) -> Result<String, Error> { ... }

    fn bool(&mut self) -> Result<bool, Error> { ... }

    fn unchecked_bool(&mut self) -> Result<bool, Error> { ... }

    fn int(&mut self) -> Result<i32, Error> { ... }

    fn unchecked_int(&mut self) -> Result<i32, Error> { ... }

    fn float(&mut self) -> Result<f32, Error> { ... }

    fn string(&mut self) -> Result<String, Error> { ... }

    fn raw(&mut self) -> Result<Vec<u8>, Error> { ... }

    fn str_to_color(&mut self, string: &str) -> Result<(u8, u8, u8, u8), Error> { ... }

    fn str_to_i32(&mut self, string: &str) -> Result<i32, Error> { ... }

    fn str_to_f32(&mut self, string: &str) -> Result<f32, Error> { ... }

    fn str_to_bool(&mut self, string: &str) -> Result<bool, Error> { ... }

    fn str_to_bytes(&mut self, string: &str) -> Result<Vec<u8>, Error> { ... }

}

Provides methods to read an Ascii archive

Provided methods

fn error(&mut self, kind: ErrorCode) -> Error

Return an Error with the given code

fn byte(&mut self) -> Result<u8, Error>

Consume a byte

fn peek(&mut self) -> Result<u8, Error>

Peek for the next byte, the own position does not change

fn peek_tuple(&mut self) -> Result<(u8, u8), Error>

Peek for the next two bytes, the own position does not change

fn test_for(&mut self, s: &str) -> Result<bool, Error>

Test if the given string matches the following string in the ascii archive

fn consume(&mut self, s: &str) -> Result<bool, Error>

Consume the given string, returns false if the following string does not match

fn consume_until(&mut self, byte: u8) -> Result<u32, Error>

Consume until the given byte occurs, returns number of bytes consumed

fn consume_whitespaces(&mut self) -> Result<(), Error>

Consumes all whitespaces until the next non whitespace char

fn string_until(&mut self, byte: u8) -> Result<String, Error>

Returns the string until the given byte

fn string_until_whitespace(&mut self) -> Result<String, Error>

Returns the string until a whitespace occurs

fn bool(&mut self) -> Result<bool, Error>

Consumes a bool string with name and kind specified and returns its value

fn unchecked_bool(&mut self) -> Result<bool, Error>

Consumes only a bool string and returns its value

fn int(&mut self) -> Result<i32, Error>

Consumes an enum or a integer with name and kind specified and returns its value

fn unchecked_int(&mut self) -> Result<i32, Error>

Consumes only a enum or int string and returns its value

fn float(&mut self) -> Result<f32, Error>

Consumes a float string with name and kind specified and returns its value

fn string(&mut self) -> Result<String, Error>

Consumes a variable of the type string with name and kind specified and returns its value

fn raw(&mut self) -> Result<Vec<u8>, Error>

Consumes a variable of the type raw vector and returns its value

fn str_to_color(&mut self, string: &str) -> Result<(u8, u8, u8, u8), Error>

Convert string to color rgba

fn str_to_i32(&mut self, string: &str) -> Result<i32, Error>

Convert string to i32

fn str_to_f32(&mut self, string: &str) -> Result<f32, Error>

Convert string to f32

fn str_to_bool(&mut self, string: &str) -> Result<bool, Error>

Convert string to bool

fn str_to_bytes(&mut self, string: &str) -> Result<Vec<u8>, Error>

Convert string to bytes

Loading content...